FinalPolish
- Refactor weapons
