+++ Assets/Game/Weapons/Resources/BowRed.asset
@@ -12,9 +12,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 20b438989a08431cafd45364a7646676, type: 3}
   m_Name: BowRed
   m_EditorClassIdentifier: 
-  equipedPrefab: {fileID: 7555281940950563468, guid: 8361cf2c161dd2a4392703a85e9c3507, type: 3}
+  equipedPrefab: {fileID: 7178779590678656227, guid: 8361cf2c161dd2a4392703a85e9c3507, type: 3}
   _animatorOveride: {fileID: 22100000, guid: 1e21fe96575cdc04c9f43ffce4e49858, type: 2}
   weaponRange: 10
   weaponDamage: 8
+  percentageModifier: 0
   isRightHanded: 0
   projectilePrefab: {fileID: 1106514011578575207, guid: 6ab317e63e2fa5e4d87ab16e9f810de7, type: 3}
