+++ Assets/Game/Weapons/Resources/Sword.asset
@@ -12,7 +12,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 20b438989a08431cafd45364a7646676, type: 3}
   m_Name: Sword
   m_EditorClassIdentifier: 
-  equipedPrefab: {fileID: 8921542051117223491, guid: 9013cc8c33f0dcd4a9c76a84ebc73ff7, type: 3}
+  equipedPrefab: {fileID: 50943672280984630, guid: 9013cc8c33f0dcd4a9c76a84ebc73ff7, type: 3}
   _animatorOveride: {fileID: 22100000, guid: 44eaa9ee89f844047be259777ea932a3, type: 2}
   weaponRange: 2
   weaponDamage: 10
+  percentageModifier: 0
+  isRightHanded: 1
+  projectilePrefab: {fileID: 0}
