+++ Assets/Scripts/Combat/Weapon.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 20b438989a08431cafd45364a7646676
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 65114187a10544098eae8ae90f843af2
+timeCreated: 1609425733
