+++ Assets/Scripts/Combat/WeaponConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20b438989a08431cafd45364a7646676
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
